requirements.txt
torch == 2.1.0
torchvision
matplotlib
torchinfo 
fastapi==0.103.2
uvicorn==0.23.2
tqdm
typing
streamlit
opencv-python
numpy 
pillow
langchain
pydantic
vertexai
google-cloud-aiplatform==1.33.0
